chore(tasks): convert SandboxProtocol to SandboxBase ABC and deduplicate clone_repository
Convert the sandbox duck-typing Protocol into a proper ABC. Move the
identical clone_repository and __enter__/__exit__ implementations from
DockerSandbox and ModalSandbox into the shared base class. Consolidate
WORKING_DIR constant. Rename SandboxProtocol → SandboxBase.

Made-with: Cursor
